refactor: move transform tests to separate files
